+++ packages/cli/src/zosjobs/-strings-/en.ts
@@ -42,9 +42,8 @@ export default {
         }
     },
     DELETE: {
-        SUMMARY: "Delete a z/OS job",
-        DESCRIPTION: "Delete a single job by job ID in OUTPUT status." +
-                     " This cancels the job if it is running and purges its output from the system.",
+        SUMMARY: "Delete a z/OS job or jobs",
+        DESCRIPTION: "Delete a single job by job ID or delete multiple jobs in OUTPUT status.",
         ACTIONS: {
             JOB: {
                 SUMMARY: "Delete a single job by job ID",